+++ integration/js/ContentShareVideoTest.js
@@ -33,6 +33,7 @@ class ContentShareVideoTest extends SdkBaseTest {
     //Turn on Content Share
     await test_window_1.runCommands(async () => await ClickContentShareVideoTestButton.executeStep(this, session, "ON"));
     await TestUtils.waitAround(5000);
+    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_1.runCommands(async () => await RosterCheck.executeStep(this, session, 3));
     await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_2.runCommands(async () => await RosterCheck.executeStep(this, session, 3));
@@ -40,8 +41,9 @@ class ContentShareVideoTest extends SdkBaseTest {
     //Turn off Content Share
     await test_window_1.runCommands(async () => await ClickContentShareButton.executeStep(this, session, "OFF"));
     await TestUtils.waitAround(5000);
+    await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "OFF", 1));
     await test_window_2.runCommands(async () => await RosterCheck.executeStep(this, session, 2));
-    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "OFF", 2));
+    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "OFF", 1));
     await test_window_1.runCommands(async () => await RosterCheck.executeStep(this, session, 2));
   }
 }